content_creation/requirements-shared.txt
# New dependencies for the shared publish/notify/config layer
# Install with: pip install -r requirements-shared.txt
#
# Do NOT pip install gradio here — the config UI (Phase 7) requires
# isolation from AnimateDiff's Gradio 3.36.1 pin.

# Config schema and validation
pydantic>=2.0,<3.0
pyyaml>=6.0

# Env management
python-dotenv>=1.0

# YouTube publishing (Phase 5)
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0

# Notifications (Phase 4)
discord-webhook>=1.3.0
slack-sdk>=3.30.0
